requirements.txt
pandas
requests
#psycopg2-binary
#sqlalchemy
psycopg2-binary